Move antispam code closer to root directory
