+++ antispam/gibberishdetector/__init__.py
